pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apab"
version = "0.2.0"
description = "Agentic Phased Array Builder — LLM-driven phased-array antenna design via MCP tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "APAB Contributors" }]
keywords = ["phased-array", "antenna", "beamforming", "mcp", "llm", "agent", "rf", "electromagnetics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "edgefem>=1.0",
    "phased-array-modeling>=1.0",
    "phased-array-systems>=0.4",
    "numpy>=1.24",
    "scipy>=1.10",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "rich>=13.0",
    "diskcache>=5.6",
    "mcp>=1.26",
    "h5py>=3.8",
]

[project.urls]
Homepage = "https://github.com/jman4162/agentic-phased-array-builder"
Repository = "https://github.com/jman4162/agentic-phased-array-builder"
Issues = "https://github.com/jman4162/agentic-phased-array-builder/issues"
Changelog = "https://github.com/jman4162/agentic-phased-array-builder/blob/main/CHANGELOG.md"

[project.optional-dependencies]
ollama = ["ollama>=0.4"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-generativeai>=0.5"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-timeout>=2.1",
    "mypy>=1.5",
    "ruff>=0.1",
    "coverage>=7.0",
    "types-PyYAML>=6.0",
    "twine>=5.0",
]

[project.scripts]
apab = "apab.cli:main"

[project.entry-points."apab.llm_providers"]
ollama = "apab.providers.ollama:OllamaProvider"
openai = "apab.providers.openai:OpenAIProvider"
anthropic = "apab.providers.anthropic:AnthropicProvider"
gemini = "apab.providers.gemini:GeminiProvider"
openai_compatible = "apab.providers.openai_compat:OpenAICompatibleProvider"

[project.entry-points."apab.compute_backends"]
local = "apab.compute.local:LocalBackend"

[project.entry-points."apab.em_adapters"]

[tool.hatch.build.targets.wheel]
packages = ["src/apab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
